Rename tools to show they are compatible with PySide2 and allow installation parallel to PySide1 tools.
