+++ CMakeLists.txt
@@ -5,12 +5,12 @@ find_package(Shiboken2 2.0.0 REQUIRED)
 find_package(Qt5 REQUIRED Core)
 find_package(PySide2 2.0.0 REQUIRED)
 
-set(pyside_tools_MAJOR_VERSION "2")
-set(pyside_tools_MINOR_VERSION "0")
-set(pyside_tools_MICRO_VERSION "0")
-set(pyside2_tools_VERSION "${pyside_tools_MAJOR_VERSION}.${pyside_tools_MINOR_VERSION}.${pyside_tools_MICRO_VERSION}")
+set(pyside2_tools_MAJOR_VERSION "2")
+set(pyside2_tools_MINOR_VERSION "0")
+set(pyside2_tools_MICRO_VERSION "0")
+set(pyside2_tools_VERSION "${pyside2_tools_MAJOR_VERSION}.${pyside2_tools_MINOR_VERSION}.${pyside2_tools_MICRO_VERSION}")
 
-configure_file("pysideuic/__init__.py.in" "__init__.py" @ONLY)
+configure_file("pyside2uic/__init__.py.in" "pyside2uic/__init__.py" @ONLY)
 
 option(BUILD_TESTS "Build tests." TRUE)
 
@@ -29,7 +29,7 @@ if (NOT SITE_PACKAGE)
     endif()
 endif()
 
-install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/pyside-uic
+install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/pyside2-uic
         DESTINATION bin
         PERMISSIONS
         OWNER_EXECUTE OWNER_WRITE OWNER_READ
@@ -37,16 +37,16 @@ install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/pyside-uic
         WORLD_EXECUTE WORLD_READ)
 
 install(FILES "${CMAKE_CURRENT_BINARY_DIR}/__init__.py"
-        DESTINATION "${SITE_PACKAGE}/pysideuic")
-install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/pysideuic
+        DESTINATION "${SITE_PACKAGE}"/pyside2uic)
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/pyside2uic
         DESTINATION ${SITE_PACKAGE}
         PATTERN "*.in" EXCLUDE
-        PATTERN "*pysideuic/__init__.py" EXCLUDE
-        PATTERN "*pysideuic\\__init__.py" EXCLUDE)
+        PATTERN "*pyside2uic/__init__.py" EXCLUDE
+        PATTERN "*pyside2uic\\__init__.py" EXCLUDE)
 
-# Man pages for pyside-uic
+# Man pages for pyside2-uic
 if (NOT win32)
-    file(GLOB manpages "${CMAKE_CURRENT_SOURCE_DIR}/pysideuic/*.1")
+    file(GLOB manpages "${CMAKE_CURRENT_SOURCE_DIR}/pyside2uic/*.1")
     install(FILES ${manpages} DESTINATION share/man/man1)
 endif()
 
@@ -57,7 +57,7 @@ configure_file("${CMAKE_CURRENT_SOURCE_DIR}/cmake_uninstall.cmake"
 add_custom_target(uninstall "${CMAKE_COMMAND}"
                   -P "${CMAKE_CURRENT_BINARY_DIR}/cmake_uninstall.cmake")
 
-set(ARCHIVE_NAME ${CMAKE_PROJECT_NAME}-${pyside_tools_VERSION})
+set(ARCHIVE_NAME ${CMAKE_PROJECT_NAME}-${pyside2_tools_VERSION})
 add_custom_target(dist
     COMMAND mkdir -p "${CMAKE_BINARY_DIR}/${ARCHIVE_NAME}" &&
             git log > "${CMAKE_BINARY_DIR}/${ARCHIVE_NAME}/ChangeLog" &&
@@ -67,7 +67,7 @@ add_custom_target(dist
             echo "Source package created at ${CMAKE_BINARY_DIR}/${ARCHIVE_NAME}.tar.bz2.\n"
     WORKING_DIRECTORY ${CMAKE_SOURCE_DIR})
 
-set(PYSIDERCC_EXECUTABLE ${CMAKE_CURRENT_BINARY_DIR}/pyrcc/pyside-rcc)
+set(PYSIDE2RCC_EXECUTABLE ${CMAKE_CURRENT_BINARY_DIR}/pyrcc/pyside2-rcc)
 add_subdirectory(pyrcc)
 add_subdirectory(pylupdate)
 
@@ -75,4 +75,3 @@ if (BUILD_TESTS)
     enable_testing()
     add_subdirectory(tests)
 endif ()
-
