+++ pylupdate/CMakeLists.txt
@@ -19,8 +19,8 @@ find_package(Qt5Widgets)
 set(lupdate_MOC_HEADERS translator.h)
 qt5_wrap_cpp(lupdate_MOC_OUTFILES ${lupdate_MOC_HEADERS})
 
-add_executable(pyside-lupdate ${lupdate_SRC} ${lupdate_MOC_OUTFILES})
-include_directories(pyside-lupdate
+add_executable(pyside2-lupdate ${lupdate_SRC} ${lupdate_MOC_OUTFILES})
+include_directories(pyside2-lupdate
                     ${CMAKE_CURRENT_SOURCE_DIR}
                     ${Qt5Xml_INCLUDE_DIRS}
                     ${Qt5Core_INCLUDE_DIRS}
@@ -28,18 +28,17 @@ include_directories(pyside-lupdate
                     ${Qt5Widgets_INCLUDE_DIRS}
                     )
 
-target_link_libraries(pyside-lupdate
+target_link_libraries(pyside2-lupdate
                       ${Qt5Core_LIBRARIES}
                       ${Qt5Xml_LIBRARIES}
                       ${Qt5Gui_LIBRARIES}
                       ${Qt5Widgets_LIBRARIES}
                       )
 
-install(TARGETS pyside-lupdate RUNTIME DESTINATION bin)
+install(TARGETS pyside2-lupdate RUNTIME DESTINATION bin)
 
 # Man pages
 if (NOT win32)
     file(GLOB manpages "${CMAKE_CURRENT_SOURCE_DIR}/*.1")
     install(FILES ${manpages} DESTINATION share/man/man1)
 endif()
-
