+++ pylupdate/pyside2-lupdate.1
@@ -1,12 +1,12 @@
-.TH PYSIDE-LUPDATE "1" "December 2010" "pyside-lupdate" "User Commands"
+.TH PYSIDE-LUPDATE "1" "December 2010" "pyside2-lupdate" "User Commands"
 .SH NAME
-pyside\-lupdate \- extracts translatable messages from Qt UI files and Python source code.
+pyside2\-lupdate \- extracts translatable messages from Qt UI files and Python source code.
 .SH DESCRIPTION
 .SS "Usage:"
 .IP
-pyside\-lupdate [options] project\-file
+pyside2\-lupdate [options] project\-file
 .IP
-pyside\-lupdate [options] source\-files \-ts ts\-files
+pyside2\-lupdate [options] source\-files \-ts ts\-files
 .SS "Options:"
 .TP
 \fB\-noobsolete
@@ -16,7 +16,7 @@ Drop all obsolete strings
 Explain what is being done
 .TP
 \fB\-version
-Display the version of pyside-lupdate and exit
+Display the version of pyside2-lupdate and exit
 .TP
 \fB\-help
 Display this information and exit
