+++ pyrcc/main.cpp
@@ -45,7 +45,7 @@ static QString resourceRoot;
 bool processResourceFile(const QStringList &filenamesIn, const QString &filenameOut, bool list)
 {
     if (verbose)
-        fprintf(stderr, "PySide resource compiler\n");
+        fprintf(stderr, "PySide2 resource compiler\n");
 
     //setup
     RCCResourceLibrary library;
@@ -91,7 +91,7 @@ bool processResourceFile(const QStringList &filenamesIn, const QString &filename
 
 int showHelp(const char *argv0, const QString &error)
 {
-    fprintf(stderr, "PySide resource compiler\n");
+    fprintf(stderr, "PySide2 resource compiler\n");
     if (!error.isEmpty())
         fprintf(stderr, "%s: %s\n", argv0, error.toLatin1().constData());
     fprintf(stderr, "Usage: %s  [options] <inputs>\n\n"
