+++ pyrcc/pyside2-rcc.1
@@ -1,10 +1,10 @@
-.TH PYSIDE\-RCC "1" "December 2010" "pyside\-rcc" "User Commands"
+.TH PYSIDE\-RCC "1" "December 2010" "pyside2\-rcc" "User Commands"
 .SH NAME
-pyside\-rcc \- PySide resource compiler
+pyside2\-rcc \- PySide2 resource compiler
 .SH DESCRIPTION
 .SS "Usage:"
 .IP
-pyside\-rcc  [options] <inputs>
+pyside2\-rcc  [options] <inputs>
 .SS "Options:"
 .TP
 \fB\-o\fR file
