+++ pyrcc/rcc.cpp
@@ -399,7 +399,7 @@ RCCResourceLibrary::writeHeader(FILE *out)
     fprintf(out, "# Resource object code\n");
     fprintf(out, "#\n");
     fprintf(out, "# Created: %s\n", QDateTime::currentDateTime().toString().toUtf8().constData());
-    fprintf(out, "#      by: The Resource Compiler for PySide (Qt v%s)\n", QT_VERSION_STR);
+    fprintf(out, "#      by: The Resource Compiler for PySide2 (Qt v%s)\n", QT_VERSION_STR);
     fprintf(out, "#\n");
     fprintf(out, "# WARNING! All changes made in this file will be lost!\n");
     fprintf(out, "\n");
