+++ pyside2uic/Compiler/__init__.py
