+++ pyside2uic/Compiler/proxy_type.py
@@ -20,7 +20,7 @@
 # Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
 # 02110-1301 USA
 
-from pysideuic.Compiler.misc import Literal, moduleMember
+from pyside2uic.Compiler.misc import Literal, moduleMember
 
 
 class ProxyType(type):
@@ -42,7 +42,7 @@ class ProxyType(type):
                 raise
 
             # Avoid a circular import.
-            from pysideuic.Compiler.qtproxies import LiteralProxyClass
+            from pyside2uic.Compiler.qtproxies import LiteralProxyClass
 
             return type(name, (LiteralProxyClass, ),
                         {"module": moduleMember(type.__getattribute__(cls, "module"),
