+++ pyside2uic/Compiler/qtproxies.py
@@ -24,15 +24,15 @@
 import sys
 import re
 
-from pysideuic.Compiler.indenter import write_code
-from pysideuic.Compiler.misc import Literal, moduleMember
+from pyside2uic.Compiler.indenter import write_code
+from pyside2uic.Compiler.misc import Literal, moduleMember
 
 if sys.hexversion >= 0x03000000:
-    from pysideuic.port_v3.proxy_base import ProxyBase
-    from pysideuic.port_v3.as_string import as_string
+    from pyside2uic.port_v3.proxy_base import ProxyBase
+    from pyside2uic.port_v3.as_string import as_string
 else:
-    from pysideuic.port_v2.proxy_base import ProxyBase
-    from pysideuic.port_v2.as_string import as_string
+    from pyside2uic.port_v2.proxy_base import ProxyBase
+    from pyside2uic.port_v2.as_string import as_string
 
 i18n_strings = []
 i18n_context = ""
