+++ pyside2uic/icon_cache.py
