+++ pyside2uic/objcreator.py
@@ -23,12 +23,12 @@
 import sys
 import os.path
 
-from pysideuic.exceptions import NoSuchWidgetError, WidgetPluginError
+from pyside2uic.exceptions import NoSuchWidgetError, WidgetPluginError
 
 if sys.hexversion >= 0x03000000:
-    from pysideuic.port_v3.load_plugin import load_plugin
+    from pyside2uic.port_v3.load_plugin import load_plugin
 else:
-    from pysideuic.port_v2.load_plugin import load_plugin
+    from pyside2uic.port_v2.load_plugin import load_plugin
 
 
 # The list of directories that are searched for widget plugins.  This is
