+++ pyside2uic/port_v2/ascii_upper.py
