+++ pyside2uic/port_v3/ascii_upper.py
