+++ pyside2uic/port_v3/proxy_base.py
@@ -20,20 +20,8 @@
 # Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
 # 02110-1301 USA
 
-from pysideuic.exceptions import WidgetPluginError
+from pyside2uic.Compiler.proxy_type import ProxyType
 
 
-def load_plugin(plugin, plugin_globals, plugin_locals):
-    """ Load the given plugin (which is an open file).  Return True if the
-    plugin was loaded, or False if it wanted to be ignored.  Raise an exception
-    if there was an error.
-    """
-
-    try:
-        exec(plugin.read(), plugin_globals, plugin_locals)
-    except ImportError:
-        return False
-    except Exception as e:
-        raise WidgetPluginError("%s: %s" % (e.__class__, str(e)))
-
-    return True
+class ProxyBase(metaclass=ProxyType):
+    pass
