+++ pyside2uic/pyside2-uic.1
@@ -1,10 +1,10 @@
-.TH PYSIDE-UIC "1" "December 2010" "pyside-uic" "User Commands"
+.TH PYSIDE-UIC "1" "December 2010" "pyside2-uic" "User Commands"
 .SH NAME
-pyside\-uic \- DESCRIPTION...
+pyside2\-uic \- DESCRIPTION...
 .SH DESCRIPTION
 .SS "Usage:"
 .IP
-pyside\-uic [options] <ui\-file>
+pyside2\-uic [options] <ui\-file>
 .SS "Options:"
 .TP
 \fB\-\-version
