+++ pyside2uic/widget-plugins/qtdeclarative.py
