+++ pyside2uic/widget-plugins/qtwebkit.py
